no more printfs

+++ srcs/ft_print_types.c
@@ -6,7 +6,7 @@
 /*   By: becastro <becastro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 20:24:39 by bena              #+#    #+#             */
-/*   Updated: 2022/05/12 22:20:33 by becastro         ###   ########.fr       */
+/*   Updated: 2022/05/12 22:38:08 by becastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ void	x(t_args *lstargs)
 {
 	char	*hexa_nbr;
 
-	hexa_nbr = ft_itoa_base(va_arg(lstargs->variatic_arg, unsigned long), 16);
+	hexa_nbr = ft_itoa_base(va_arg(lstargs->variatic_arg, int), 16);
 	lstargs->printable_arg = hexa_nbr;
 }
 
@@ -48,7 +48,7 @@ void	upper_x(t_args *lstargs)
 	int		i;
 
 	i = -1;
-	hexa_nbr = ft_itoa_base(va_arg(lstargs->variatic_arg, unsigned int), 16);
+	hexa_nbr = ft_itoa_base(va_arg(lstargs->variatic_arg, int), 16);
 	while (hexa_nbr[++i])
 		hexa_nbr[i] = ft_toupper(hexa_nbr[i]);
 	lstargs->printable_arg = hexa_nbr;
