+++ srcs/ft_print_types2.c
@@ -6,7 +6,7 @@
 /*   By: becastro <becastro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/11 06:12:13 by bena              #+#    #+#             */
-/*   Updated: 2022/05/12 22:19:59 by becastro         ###   ########.fr       */
+/*   Updated: 2022/05/12 23:22:26 by becastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ void	u(t_args *lstargs)
 {
 	unsigned int	nbr;
 
-	nbr = va_arg(lstargs->variatic_arg, long long);
+	nbr = va_arg(lstargs->variatic_arg, unsigned int);
 	 if (nbr < 0)
 	 	nbr *= -1;
 	//printf("test %d\n", nbr);
